+++ src/Akeneo/AssetManager/front/application/component/app/user-navigation.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import {PimView} from '@akeneo-pim-community/shared';
+
+const UserNavigation = () => (
+  <PimView
+    className="AknTitleContainer-userMenuContainer AknTitleContainer-userMenu"
+    viewName="pim-asset-family-index-user-navigation"
+  />
+);
+
+export {UserNavigation};