+++ src/Akeneo/AssetManager/front/application/component/asset/list/filter/option.tsx
@@ -1,15 +1,17 @@
-import React from 'react';
+import React, {memo} from 'react';
+import styled from 'styled-components';
+import {Dropdown, MultiSelectInput, SwitcherButton, useBooleanState} from 'akeneo-design-system';
 import {useTranslate, getLabel} from '@akeneo-pim-community/shared';
 import {FilterView, FilterViewProps} from 'akeneoassetmanager/application/configuration/value';
 import {isOptionAttribute} from 'akeneoassetmanager/domain/model/attribute/type/option';
-import {Option, getOptionLabel} from 'akeneoassetmanager/domain/model/attribute/type/option/option';
-import Select2 from 'akeneoassetmanager/application/component/app/select2';
+import {getOptionLabel} from 'akeneoassetmanager/domain/model/attribute/type/option/option';
 import {isOptionCollectionAttribute} from 'akeneoassetmanager/domain/model/attribute/type/option-collection';
 import {getAttributeFilterKey} from 'akeneoassetmanager/tools/filter';
 import OptionCode from 'akeneoassetmanager/domain/model/attribute/type/option/option-code';
 
-const memo = (React as any).memo;
-const useState = (React as any).useState;
+const DropdownContent = styled.div`
+  padding: 5px 20px;
+`;
 
 type OptionFilterViewProps = FilterViewProps & {
   context: {
@@ -19,112 +21,57 @@ type OptionFilterViewProps = FilterViewProps & {
 
 const DEFAULT_OPERATOR = 'IN';
 
-//TODO Use DSM Multiselect?
 const OptionFilterView: FilterView = memo(({attribute, filter, onFilterUpdated, context}: OptionFilterViewProps) => {
   const translate = useTranslate();
+  const [isOpen, open, close] = useBooleanState();
+
   if (!(isOptionAttribute(attribute) || isOptionCollectionAttribute(attribute))) {
     return null;
   }
 
-  const [isOpen, setIsOpen] = useState(false);
-
-  const availableOptions = attribute.options.reduce(
-    (availableOptions: {[choiceValue: string]: string}, option: Option) => {
-      const normalizedOption: Option = option;
-      availableOptions[normalizedOption.code] = getOptionLabel(option, context.locale);
-
-      return availableOptions;
-    },
-    {} as {[label: string]: string}
-  );
-
-  const emptyFilter = () => {
-    setIsOpen(false);
-    onFilterUpdated({
-      field: getAttributeFilterKey(attribute),
-      operator: DEFAULT_OPERATOR,
-      value: [],
-      context: {},
-    });
-  };
   const value = undefined !== filter ? filter.value : [];
-  const labels = value.map((optionCode: OptionCode) =>
-    undefined !== availableOptions[optionCode] ? availableOptions[optionCode] : `[${optionCode}]`
-  );
-  const [position, setPosition] = React.useState({top: 0, left: 0});
-  const labelRef = React.useRef<HTMLSpanElement>(null);
-  const openPanel = () => {
-    setIsOpen(true);
-    if (null !== labelRef.current) {
-      const viewportOffset = labelRef.current.getBoundingClientRect();
-      setPosition({top: viewportOffset.top, left: viewportOffset.left});
-    }
-  };
+  const labels = value.map((optionCode: OptionCode) => {
+    const option = attribute.options.find(({code}) => code === optionCode) ?? null;
+
+    return null === option ? `[${optionCode}]` : getOptionLabel(option, context.locale);
+  });
+  const attributeLabel = getLabel(attribute.labels, context.locale, attribute.code);
 
   return (
-    <>
-      <span ref={labelRef} className="AknFilterBox-filterLabel" onClick={openPanel}>
-        {getLabel(attribute.labels, context.locale, attribute.code)}
-      </span>
-      <span className="AknFilterBox-filterCriteria AknFilterBox-filterCriteria--limited" onClick={openPanel}>
-        <span className="AknFilterBox-filterCriteriaHint">
-          {0 === labels.length ? translate('pim_asset_manager.asset.grid.filter.option.all') : labels.join(', ')}
-        </span>
-        <span className="AknFilterBox-filterCaret" />
-      </span>
-      {isOpen ? (
-        <div>
-          <div className="AknDropdown-mask" onClick={() => setIsOpen(false)} />
-          <div
-            className="AknFilterBox-filterDetails AknFilterBox-filterDetails--rightAlign"
-            style={{top: `${position.top + 20}px`, left: `${position.left}px`, position: 'fixed'}}
-          >
-            <div className="AknFilterChoice">
-              <div className="AknFilterChoice-header">
-                <div className="AknFilterChoice-title">
-                  {getLabel(attribute.labels, context.locale, attribute.code)}
-                </div>
-                <div className="AknIconButton AknIconButton--erase" onClick={emptyFilter} />
-              </div>
-              <div>
-                <Select2
-                  className="asset-option-selector"
-                  data={availableOptions}
-                  value={value}
-                  multiple={true}
-                  readOnly={false}
-                  configuration={{
-                    allowClear: true,
-                    placeholder: translate('pim_asset_manager.asset.grid.filter.option.no_value'),
-                    formatSelection: (data: any, _container: any, escapeMarkup: any) => {
-                      const result = data ? escapeMarkup(data.text) : undefined;
-                      if (result !== undefined) {
-                        return '<div title="' + result + '">' + result + '</div>';
-                      }
-
-                      return result;
-                    },
-                    formatResult: (result: any, container: any, _query: any, _escapeMarkup: any) => {
-                      const formerResult = '<span class="select2-match"></span>' + result.text;
-                      container.attr('title', result.text);
-                      return formerResult;
-                    },
-                  }}
-                  onChange={(optionCodes: string[]) => {
-                    onFilterUpdated({
-                      field: getAttributeFilterKey(attribute),
-                      operator: DEFAULT_OPERATOR,
-                      value: optionCodes,
-                      context: {},
-                    });
-                  }}
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-      ) : null}
-    </>
+    <Dropdown>
+      <SwitcherButton inline={false} label={attributeLabel} onClick={open}>
+        {0 === labels.length ? translate('pim_asset_manager.asset.grid.filter.option.all') : labels.join(', ')}
+      </SwitcherButton>
+      {isOpen && (
+        <Dropdown.Overlay verticalPosition="down" onClose={close}>
+          <Dropdown.Header>
+            <Dropdown.Title>{attributeLabel}</Dropdown.Title>
+          </Dropdown.Header>
+          <DropdownContent>
+            <MultiSelectInput
+              openLabel={translate('pim_common.open')}
+              emptyResultLabel={translate('pim_common.no_result')}
+              removeLabel={translate('pim_common.remove')}
+              value={value}
+              onChange={optionCodes =>
+                onFilterUpdated({
+                  field: getAttributeFilterKey(attribute),
+                  operator: DEFAULT_OPERATOR,
+                  value: optionCodes,
+                  context: {},
+                })
+              }
+            >
+              {attribute.options.map(option => (
+                <MultiSelectInput.Option key={option.code} value={option.code}>
+                  {getOptionLabel(option, context.locale)}
+                </MultiSelectInput.Option>
+              ))}
+            </MultiSelectInput>
+          </DropdownContent>
+        </Dropdown.Overlay>
+      )}
+    </Dropdown>
   );
 });