+++ CHANGELOG.md
@@ -133,6 +133,7 @@
 - PIM-10203: Fix product variants are not all uncategorized after a mass "Remove from categories" action
 - PIM-10211: Flatten image layers during thumbnail and preview transformations in AM
 - PIM-10216: Fix PDF renderer in case of empty tables
+- PIM-10221: Fix media link input caret jumping when manually typing a link
 
 ## Improvements